webmail/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
Flask-Session==0.5.0
Flask-Cors==4.0.0
WTForms==3.1.1
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
IMAPClient==3.0.1
email-validator==2.1.0
bleach==6.1.0
Markdown==3.5.1
python-dateutil==2.8.2
redis==5.0.1
Pillow==10.1.0
ua-parser==0.18.0
user-agents==2.2.0
